added custom authentication/aythorization

+++ src/main/java/com/example/springbeginner/services/BeerServiceImpl.java
@@ -2,7 +2,9 @@ package com.example.springbeginner.services;
 
 import com.example.springbeginner.exceptions.DuplicateEntityException;
 import com.example.springbeginner.exceptions.EntityNotFoundException;
+import com.example.springbeginner.exceptions.UnauthorizedOperationException;
 import com.example.springbeginner.models.Beer;
+import com.example.springbeginner.models.User;
 import com.example.springbeginner.repositories.BeerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -30,7 +32,7 @@ public class BeerServiceImpl implements BeerService {
     }
 
     @Override
-    public void create(Beer beer) {
+    public void create(Beer beer, User user) {
         boolean duplicateExists = true;
         try {
             repository.get(beer.getName());
@@ -46,7 +48,10 @@ public class BeerServiceImpl implements BeerService {
     }
 
     @Override
-    public void update(Beer beer) {
+    public void update(Beer beer, User user) {
+        if(!user.isAdmin()){
+            throw new UnauthorizedOperationException("Only admins can modify beer.");
+        }
         boolean duplicateExists = true;
         try {
             Beer existingBeer = repository.get(beer.getName());
@@ -65,7 +70,10 @@ public class BeerServiceImpl implements BeerService {
     }
 
     @Override
-    public void delete(int id) {
+    public void delete(int id, User user) {
+        if(!user.isAdmin()){
+            throw new UnauthorizedOperationException("Only admins can delete beer");
+        }
         repository.delete(id);
     }
 }
